Add index.ts and util.ts. Only export gs2vl()
