+++ src-ext/Typescript/ggvega/rollup.config.js
@@ -3,7 +3,7 @@ import nodeResolve from 'rollup-plugin-node-resolve';
 import json from 'rollup-plugin-json';
 
 export default {
-  input: 'build/ggvega/src/compile.js',
+  input: 'build/ggvega/src/index.js',
   output: {
     file: 'build/ggvega.js',
     format: 'umd',
