+++ src-ext/Typescript/ggvega/src/index.ts
@@ -1,0 +1,1 @@
+export {gs2vl} from './compile';
