+++ src-ext/Typescript/ggvega/test/compile.test.ts
@@ -1,4 +1,4 @@
-import {gs2vl, TranslateDatasets, TranslateLayers, removeEmpty} from '../src/compile';
+import {gs2vl, TranslateDatasets, TranslateLayers} from '../src/compile';
 import * as ggSpec from './ggSpec';
 import * as vlSpec from './vlSpec';
 
@@ -64,34 +64,5 @@ describe('compile/TranslateLayers', () => {
     );
 
     expect(vlLayers).toEqual(vlSpec.vlSpec01['layer']);
-  });
-});
-
-describe('compile/removeEmpty', () => {
-  it('should omit all empty objects in an object', () => {
-    const empty01 = {};
-
-    removeEmpty(empty01);
-
-    expect(empty01).toEqual({});
-
-    const empty02 = {
-      data: {
-        metadata: null,
-        values: [{a: 'A', b: 28}],
-        observations: {
-          values: {}
-        }
-      },
-      title: undefined,
-      layer: {},
-      scale: {
-        class: null
-      }
-    };
-
-    removeEmpty(empty02);
-
-    expect(empty02).toEqual({data: {values: [{a: 'A', b: 28}]}});
   });
 });
