+++ src-ext/Typescript/ggvega/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "files": ["src/compile.ts"],
+  "files": ["src/index.ts"],
   "compilerOptions": {
     /* Basic Options */
     // "incremental": true,                   /* Enable incremental compilation */
